requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
flask-cors==4.0.0
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4